+++ R/02_seu_batch_correct.R
@@ -1,0 +1,441 @@
+# Normalization ----
+
+#' Batch-wise Normalization of Data Using Seurat
+#'
+#' This function normalizes the data in a batch-wise manner using Seurat.
+#' The purpose is to minimize the impact of batch effects when clustering the data prior to batch correction.
+#' Three normalization methods are implemented: Z-score, Rank, and Quantile normalization.
+#' Z-score is recommended for batches from a single study/experiment.
+#' Rank is recommended for data from different studies/experiments.
+#' Quantile is not recommended.
+#'
+#' @param object A Seurat object
+#' @param markers A vector of marker genes to normalize. Defaults to all genes if NULL.
+#' @param norm_method Normalization method: "scale" (Z-score), "rank", or "qnorm" (Quantile normalization). Defaults to "scale".
+#' @param ties.method Method for handling ties in rank normalization. Options are "average", "first", "last", "random", "max", or "min". Defaults to "average".
+#'
+#' @return A Seurat object with normalized data.
+#' @export
+normalize_seurat <- function(object,
+                             markers = NULL,
+                             norm_method = "scale",
+                             ties.method = "average",
+                             mc.cores = parallel::detectCores() - 1) {
+
+  # Remove case-sensitivity
+  norm_method <- tolower(norm_method)
+  ties.method <- tolower(ties.method)
+
+  # Error check
+  if (norm_method == "rank" && !ties.method %in% c("average", "first", "last", "random", "max", "min")) {
+    stop("When using norm_method = 'rank', please use an available ties.method (average, first, last, random, max, or min).")
+  }
+
+  if (is.null(markers)) {
+    warning("All rows will be normalized. This can take a while.")
+    # Get markers
+    markers <- rownames(object)
+  }
+
+  # Messaging
+  switch(norm_method,
+         "rank" = message("Ranking expression data.."),
+         "scale" = {
+           message("Scaling expression data..")
+           object <- ScaleData(object, features = markers, split.by = "batch")
+           return(object)
+           },
+         "CLR" = {
+           message("CLR normalizing expression data..")
+           object <- NormalizeData(object, normalization.method = "CLR", margin = 2)
+           return(object)
+         },
+         "qnorm" = {
+           message("Quantile normalizing expression data..")
+           object <- quantile_norm_seurat(object, markers = markers)
+           return(object)
+         },
+         stop("Please use either 'scale', 'rank', or 'qnorm' as normalization method.")
+  )
+
+  # Rank at marker positions individually for every batch
+  batches <- unique(object$batch)
+  ranked_data <- pbmcapply::pbmclapply(
+    setNames(batches, batches), function(b) {
+      batch_cells <- WhichCells(object, expression = batch == b)
+      data <- LayerData(object, "counts")[, batch_cells]
+
+      rowzeros <- rowSums(data[markers, ]) == 0
+      # if (any(rowzeros)) {
+      #   warning("A marker is 0 for an entire batch. This marker is removed.")
+      #   data <- data[!rowzeros, ]
+      # }
+      data[markers, ] <- apply(data[markers, ], 1, rank, ties.method = ties.method) / ncol(data)
+    },
+    mc.cores = mc.cores)
+  ranked_data <- do.call(rbind, ranked_data)
+
+  LayerData(object, "scale.data") <- ranked_data
+
+  return(object)
+}
+
+# Batch-wise quantile normalization per marker using Seurat
+
+quantile_norm_seurat <- function(object, markers = NULL, mc.cores = parallel::detectCores()) {
+  message("Quantile normalizing expression data..")
+  if (is.null(markers)) {
+    markers <- rownames(object)
+  }
+
+  # Determine goal distributions for each marker by getting quantiles across all batches
+  refq <- pbmclapply(setNames(markers, markers), function(m) {
+    quantile(LayerData(object, "data")[m, ], probs = seq(0, 1, length.out = 5), names = FALSE)
+  }, mc.cores = mc.cores)
+
+  for (batch in unique(object$batch)) {
+    batch_cells <- WhichCells(object, ident = batch)
+    data <- LayerData(object, "data")[, batch_cells]
+    for (m in markers) {
+      qx <- quantile(data[m, ], probs = seq(0, 1, length.out = 5), names = FALSE)
+      spf <- splinefun(x = qx, y = refq[[m]], method = "monoH.FC", ties = min)
+
+      # Apply the spline function to adjust quantiles
+      data[m, ] <- spf(data[m, ])
+    }
+    LayerData(object, "scale.data") <- data
+  }
+
+  return(object)
+}
+
+# Clustering ----
+
+#' Create Self-Organizing Map
+#'
+#' The function uses the kohonen package to create a Self-Organizing Map (SOM).
+#' It is used to segregate the cells for batch correction to make the correction less affected
+#' by samples with high abundances of a particular cell type.
+#'
+#' @param object A Seurat object
+#' @param markers A vector of marker genes to use for the SOM. Defaults to all genes if NULL.
+#' @param seed The seed to use when creating the SOM. Defaults to 473.
+#' @param xdim The x-dimension size of the SOM. Defaults to 8.
+#' @param ydim The y-dimension size of the SOM. Defaults to 8.
+#' @param rlen Number of times the data is presented to the SOM network. Defaults to 10.
+#'
+#' @return A vector of clustering labels
+#' @export
+create_som_seurat <- function(
+    object,
+    markers = NULL,
+    seed = 473,
+    rlen = 10,
+    xdim = 8,
+    ydim = 8) {
+
+  # Default to all genes if markers are not specified
+  if (is.null(markers)) {
+    markers <- rownames(object)
+  }
+
+  # Extract data for the markers
+  data <- LayerData(object, "scale.data")[markers, ]
+
+  # SOM grid on overlapping markers, extract clustering per cell
+  message("Creating SOM grid..")
+  set.seed(seed)
+  som_grid <- kohonen::somgrid(xdim = xdim, ydim = ydim, topo = "rectangular")
+  som_model <- kohonen::som(data, grid = som_grid, rlen = rlen, dist.fcts = "euclidean")
+
+  # Add labels to metadata
+  object <- AddMetaData(object, metadata = som_model$unit.classif, col.name = "Labels")
+
+  return(object)
+}
+
+
+# Batch correction ----
+
+
+
+#' Correct data using ComBat
+#'
+#' Compute the batch correction on the data using the ComBat algorithm.
+#' Define a covariate, either as a character vector or name of Seurat metadata column.
+#' The covariate should preferably be the cell condition types but can be any column that infers heterogeneity in the data.
+#' The function assumes that the batch information is in the "batch" column and the data contains a "sample" column with sample information.
+#'
+#' @param object A Seurat object.
+#' @param label The cluster or cell type label. Either as a metadata column name or vector.
+#' @param markers A vector of marker genes to use for the correction. Defaults to all genes if NULL.
+#' @param method The method for batch correction. Choose "ComBat" for cytometry data and "ComBat_seq" for bulk RNAseq data. Defaults to "ComBat".
+#' @param covar The covariate ComBat should use. Can be a vector or a metadata column name in the input Seurat object. If NULL, no covar will be used.
+#' @param anchor A column or vector specifying which samples are replicates and which are not. If specified, this column will be used as a covariate in ComBat. Be aware that it may be confounded with the condition.
+#' @param ref.batch Optional. A string of the batch that should be used as the reference for batch adjustments.
+#' @param parametric Logical. If TRUE, the parametric version of ComBat is used. If FALSE, the non-parametric version is used. Defaults to TRUE.
+#'
+#' @return A Seurat object with corrected data.
+#' @export
+correct_data_seurat <- function(
+    object,
+    markers = NULL,
+    mc.cores = 1,
+    method = c("ComBat", "ComBat_seq"),
+    covar = NULL,
+    anchor = NULL,
+    ref.batch = NULL,
+    parametric = TRUE,
+    return_seurat = TRUE) {
+
+
+  method <- match.arg(method)
+  message("Batch correcting data..")
+
+  metadata <- object[[]]
+
+  # Check for batch column
+  if (!"batch" %in% colnames(metadata)) {
+    stop("The 'batch' column is missing in the metadata.")
+  }
+
+  if (is.null(markers)) {
+    markers <- rownames(object)
+  }
+
+
+  # Add covar to metadata if it's a vector
+  if (!is.null(covar) && length(covar) > 1) {
+    object$covar <- covar
+    covar <- "covar"
+  } else if (length(covar) == 1) {
+    stopifnot("The covar column is missing" = covar %in% colnames(metadata))
+  }
+
+  # Add anchor to metadata if it's a vector
+  if (!is.null(anchor) && length(anchor) > 1) {
+    object$anchor <- anchor
+    anchor <- "anchor"
+  } else if (length(anchor) == 1) {
+    stopifnot("The anchor column is missing" = anchor %in% colnames(metadata))
+  }
+
+
+  message("Batch correcting..")
+  labels <- unique(object$Labels)
+  corrected_data <- lapply(
+    setNames(labels, labels),
+    function(lab) {
+      label_cells <- WhichCells(object, expression = Labels == lab)
+      object_lab <- object[markers, label_cells]
+      correct_label(
+        object_lab,
+        covar = covar,
+        anchor = anchor,
+        parametric = parametric,
+        ref.batch = ref.batch,
+        method = method
+      )}
+  )
+
+  corrected_data <- do.call(cbind, corrected_data)
+
+  # Ensure data is in the same order as the original Seurat object
+  corrected_data <- corrected_data[, match(colnames(object), colnames(corrected_data))]
+
+  if (!return_seurat) return(corrected_data)
+  # Update the Seurat object with corrected data
+
+  object[["cyCombine"]] <- CreateAssayObject(data = corrected_data, key = "corrected.data_")
+
+
+  return(object)
+}
+
+# Function to perform ComBat correction
+combat <- function(object_lab, mod_matrix, parametric, ref.batch, method) {
+
+  if (method == "ComBat") {
+    data <- ComBat(
+      dat = as.matrix(LayerData(object_lab, layer = "data")),
+      batch = as.character(object_lab$batch),
+      mod = mod_matrix,
+      par.prior = parametric,
+      ref.batch = ref.batch,
+      prior.plots = FALSE
+    )
+  } else if (method == "ComBat_seq") {
+    data <- ComBat_seq(
+      counts = as.matrix(LayerData(object_lab, "counts")),
+      batch = as.character(object_lab$batch),
+      covar_mod = mod_matrix,
+      full_mod = TRUE
+    )
+  }
+  return(data)
+}
+
+# Function to correct each group
+correct_label <- function(object_lab, covar, anchor, parametric, ref.batch, method) {
+  num_covar <- 1
+  num_anchor <- 1
+  num_batches <- nlevels(factor(object_lab$batch))
+
+  if (num_batches == 1) {
+    message(paste("Label group", object_lab$Labels[1], "only contains cells from batch", object_lab$batch[1]))
+    return(data)
+  }
+
+  if (!is.null(covar) && !check_confound(object_lab$batch, object_lab[[covar]][,1])) {
+    num_covar <- nlevels(factor(object_lab[[covar]][,1]))
+    if (num_covar == 1) covar <- NULL
+  } else {
+    covar <- NULL
+  }
+
+  if (!is.null(anchor) && !check_confound(object_lab$batch, object_lab[[anchor]][,1])) {
+    num_anchor <- nlevels(factor(object_lab[[anchor]][,1]))
+    if (num_anchor == 1) anchor <- NULL
+  } else {
+    anchor <- NULL
+  }
+
+  if (num_covar > 1 && num_anchor > 1 && check_confound(object_lab$batch, interaction(object_lab[[covar]][,1], object_lab[[anchor]][,1]))) {
+    anchor <- NULL
+  }
+
+  if (!is.null(covar) && !is.null(anchor)) {
+    mod_matrix <- model.matrix(~ object_lab[[covar]][,1] + object_lab[[anchor]][,1])
+  } else if (!is.null(covar)) {
+    mod_matrix <- model.matrix(~ object_lab[[covar]][,1])
+  } else if (!is.null(anchor)) {
+    mod_matrix <- model.matrix(~ object_lab[[anchor]][,1])
+  } else {
+    mod_matrix <- NULL
+  }
+
+  data <- combat(object_lab, mod_matrix, parametric, ref.batch, method)
+
+  return(data)
+}
+
+# Function to check confounding
+check_confound <- function(batch, covariate) {
+  covariate_levels <- unique(covariate)
+  for (level in covariate_levels) {
+    if (length(unique(batch[covariate == level])) == 1) {
+      return(TRUE)
+    }
+  }
+  return(FALSE)
+}
+
+# Wrapper ----
+
+#' Run batch correction on data
+#'
+#' This is a wrapper function for the cyCombine batch correction workflow.
+#'  To run the workflow manually, type "batch_correct" to see the source code of this wrapper and follow along or read the vignettes on the GitHub page \url{https://github.com/biosurf/cyCombine}.
+#'
+#' @inheritParams create_som
+#' @inheritParams correct_data
+#' @inheritParams normalize
+#' @family batch
+#' @examples
+#' \dontrun{
+#' corrected <- uncorrected %>%
+#'   batch_correct(markers = markers,
+#'   covar = "condition")
+#'   }
+#' @export
+batch_correct_seurat <- function(
+    object,
+    xdim = 8,
+    ydim = 8,
+    rlen = 10,
+    parametric = TRUE,
+    method = c("ComBat", "ComBat_seq"),
+    ref.batch = NULL,
+    seed = 473,
+    covar = NULL,
+    anchor = NULL,
+    markers = NULL,
+    norm_method = "scale",
+    ties.method = "average",
+    mc.cores = parallel::detectCores() - 1) {
+
+  stopifnot(
+    "No 'batch' column in data." = "batch" %in% names(object[[]]))
+
+  layer <- switch(
+    norm_method,
+    "scale" = "scale.data",
+    "rank" = "rank.data")
+
+  if (is.null(markers)) {
+    markers <- rownames(object)
+  }
+
+  object <- object[markers, ]
+
+  for (i in seq_len(max(length(xdim), length(ydim)))) {
+    xdim_i <- xdim[min(length(xdim), i)]
+    ydim_i <- ydim[min(length(ydim), i)]
+
+    message("Batch correcting using a SOM grid of dimensions ", xdim_i,"x", ydim_i)
+
+    # Create SOM on normalized data
+    object <- normalize_seurat(
+      object,
+      markers = markers,
+      norm_method = norm_method,
+      ties.method = ties.method,
+      mc.cores = mc.cores)
+    object <- create_som_seurat(
+      object,
+      markers = markers,
+      rlen = rlen,
+      seed = seed,
+      xdim = xdim_i,
+      ydim = ydim_i)
+
+
+
+    # Run batch correction
+    labels <- unique(object$Labels)
+    corrected_data <- pbmclapply(
+      setNames(labels, labels),
+      function(lab) {
+        label_cells <- WhichCells(object, expression = Labels == lab)
+        object_lab <- object[markers, label_cells]
+        correct_data_seurat(
+          object_lab,
+          covar = covar,
+          anchor = anchor,
+          markers = markers,
+          parametric = parametric,
+          method = method,
+          ref.batch = ref.batch,
+          mc.cores = 1,
+          return_seurat = FALSE
+        )
+      },
+      mc.cores = mc.cores
+    )
+  }
+
+  corrected_data <- do.call(cbind, corrected_data)
+
+  # Ensure data is in the same order as the original Seurat object
+  corrected_data <- corrected_data[, match(colnames(object), colnames(corrected_data))]
+
+  object[["cyCombine"]] <- CreateAssayObject(data = corrected_data, key = "cycombine_")
+
+  DefaultAssay(object) <- "cyCombine"
+
+  message("Done!")
+  return(object)
+}
+
+
+